EmployeeService/requirements.txt
nameko
mysql-connector-python
bcrypt
pyjwt
dotenv
